requirements.txt
Flask==0.12.2
Flask-PyMongo==0.5.1
gunicorn==19.7.1
flask-bcrypt==0.7.1